Build 9601
Calculator shows loading hourglass when using long method for exponents.

Prevents too many explorer windows opening, fix bugs with desktop icons. Update calculator to clear display, then the stack. Fixed major bugs with JohnOS Update that caused most of the lag in the project. Updated the benchmark program to run the benchmark for 3x as long and average for more accurate results. Removed dependence on the timer because clicking the mouse resets the timer. Also, performance is way better than before so I adjusted the bounds of the test. Scores from the previous version are still comparable to the new score.